CI cleanup: fmt, clippy hard-errors, LoC-cap splits
- Remove leftover PAIR_DUMP debug diagnostic from verify_dflash_step.rs
  (same class as already-removed ITER/bisect/FORCE_DECODE_SEQLEN_RANGE
  tooling).
- cargo fmt across files touched by the branch.
- Fix two clippy::collapsible_if / unnecessary_unwrap hard errors
  (#![deny(clippy::all)] in spark-runtime and spark-model) that were
  blocking cargo clippy --workspace --tests: nvtx_diag.rs's Drop impl,
  and trait_prefill_phase1.rs's qkvz_fix branch.
- Split trait_prefill_gdn.rs (517 lines) by extracting the Q12 Path B
  batched-recurrence dispatch into trait_prefill_gdn_batched.rs.
- Split verify_d.rs (508 lines) by extracting K=gamma attention-metadata
  upload into verify_d_kgamma_attn_meta.rs (build_kgamma_attn_metadata
  takes block_size as a parameter rather than re-locking kv_cache, since
  the caller already holds that non-reentrant lock).

Verified after these fixes: cargo fmt --check clean, cargo clippy
--workspace --tests clean (spark-runtime, spark-model, spark-server all
pass with no newly-surfaced errors), full cargo clean + release rebuild
with 0 warnings/errors.
