+++ crates/spark-model/src/layers/qwen3_ssm/trait_prefill_gdn.rs
@@ -363,155 +363,4 @@ impl Qwen3SsmLayer {
 
         Ok(())
     }
-
-    /// Q12 Path B: batched GDN recurrence — mirrors prefill_gdn_full_inner
-    /// dispatch ladder but routes to the `*_batched` kernel variants and
-    /// passes `h_state_ptrs` (device array of N pointers) instead of a
-    /// single h_state device pointer.
-    ///
-    /// Constraint: scheduler-enforced same-chunk-len across all N streams.
-    /// `gdn_bufs.qkv` / `gate_beta` / `output` are stacked
-    /// `[batch_size, chunk_len, *]` contiguous in memory. Each batch
-    /// element's QKV starts at `b * chunk_len * conv_dim` (BF16).
-    ///
-    /// Validation status: kernels unvalidated against hardware.
-    pub(super) fn prefill_gdn_full_batched_inner(
-        &self,
-        h_state_ptrs: spark_runtime::gpu::DevicePtr,
-        gdn_bufs: &GdnPrefillBuffers,
-        batch_size: u32,
-        chunk_len: u32,
-        ctx: &ForwardContext,
-        stream: u64,
-    ) -> Result<()> {
-        let nk = ctx.config.linear_num_key_heads;
-        let kd = ctx.config.linear_key_head_dim;
-        let nv = ctx.config.linear_num_value_heads;
-        let vd = ctx.config.linear_value_head_dim;
-        let key_dim = nk * kd;
-        let value_dim = nv * vd;
-        let conv_dim = key_dim * 2 + value_dim;
-        let bf16 = 2usize;
-        let fp32 = 4usize;
-
-        let q_ptr = gdn_bufs.qkv;
-        let k_ptr = gdn_bufs.qkv.offset(key_dim * bf16);
-        let v_ptr = gdn_bufs.qkv.offset(key_dim * 2 * bf16);
-        let gate_ptr = gdn_bufs.gate_beta;
-        let beta_ptr = gdn_bufs.gate_beta.offset(nv * fp32);
-        let gb_stride = (nv * 2) as u32;
-
-        // Mirror the single-stream dispatch ladder. Total tokens per stream
-        // is `chunk_len`; the kernel internally processes `batch_size` such
-        // streams (grid dim Y).
-        if self.gdn_prefill_wy32_batched_k.0 != 0 && chunk_len > 32 {
-            // #110: dynamic smem must cover the FULL kernel layout (H + smem_k +
-            // smem_q + smem_warp[4] + smem_kd[C*C] + smem_g[C] + smem_bt[C], C=32).
-            // The old `+256` slack under-counted the smem_warp(16)+smem_g(128)+
-            // smem_bt(128)=272 trailer by 16 B, so the kernel's smem_bt tail wrote
-            // past the requested allocation → CUDA illegal access under live
-            // occupancy (compute-sanitizer: Invalid __shared__ write at +0xce0).
-            let smem =
-                (kd * vd * 4 + 32 * kd * 2 + 32 * kd * 2 + 32 * 32 * 4 + (4 + 32 + 32) * 4) as u32;
-            ops::gdn_prefill_persistent_smem_batched(
-                ctx.gpu,
-                self.gdn_prefill_wy32_batched_k,
-                h_state_ptrs,
-                q_ptr,
-                k_ptr,
-                v_ptr,
-                gate_ptr,
-                beta_ptr,
-                gdn_bufs.output,
-                batch_size,
-                chunk_len,
-                nk as u32,
-                nv as u32,
-                kd as u32,
-                vd as u32,
-                conv_dim as u32,
-                conv_dim as u32,
-                gb_stride,
-                smem,
-                stream,
-            )?;
-        } else if self.gdn_prefill_persistent_wy4_batched_k.0 != 0 {
-            let smem = (kd * vd * 4 + 8 * kd * 4 + 56) as u32;
-            ops::gdn_prefill_persistent_smem_batched(
-                ctx.gpu,
-                self.gdn_prefill_persistent_wy4_batched_k,
-                h_state_ptrs,
-                q_ptr,
-                k_ptr,
-                v_ptr,
-                gate_ptr,
-                beta_ptr,
-                gdn_bufs.output,
-                batch_size,
-                chunk_len,
-                nk as u32,
-                nv as u32,
-                kd as u32,
-                vd as u32,
-                conv_dim as u32,
-                conv_dim as u32,
-                gb_stride,
-                smem,
-                stream,
-            )?;
-        } else if (256..=4096).contains(&chunk_len) && self.gdn_prefill_persistent_batched_k.0 != 0
-        {
-            ops::gdn_prefill_persistent_batched(
-                ctx.gpu,
-                self.gdn_prefill_persistent_batched_k,
-                h_state_ptrs,
-                q_ptr,
-                k_ptr,
-                v_ptr,
-                gate_ptr,
-                beta_ptr,
-                gdn_bufs.output,
-                batch_size,
-                chunk_len,
-                nk as u32,
-                nv as u32,
-                kd as u32,
-                vd as u32,
-                conv_dim as u32,
-                conv_dim as u32,
-                gb_stride,
-                stream,
-            )?;
-        } else if self.gdn_prefill_split4_batched_k.0 != 0 {
-            ops::gdn_prefill_split4_batched(
-                ctx.gpu,
-                self.gdn_prefill_split4_batched_k,
-                h_state_ptrs,
-                q_ptr,
-                k_ptr,
-                v_ptr,
-                gate_ptr,
-                beta_ptr,
-                gdn_bufs.output,
-                batch_size,
-                chunk_len,
-                nk as u32,
-                nv as u32,
-                kd as u32,
-                vd as u32,
-                conv_dim as u32,
-                conv_dim as u32,
-                gb_stride,
-                stream,
-            )?;
-        } else {
-            anyhow::bail!(
-                "Qwen3SsmLayer::prefill_gdn_full_batched_inner: no batched GDN \
-                 kernel handle is loaded for this target — caller should fall \
-                 back to per-stream prefill_gdn_full."
-            );
-        }
-
-        Ok(())
-    }
 }
